refactor: move rebase functions out of virtual_branches module
This opens the doors for further simplifying the core
