+++ crates/gitbutler-core/src/lib.rs
@@ -27,6 +27,7 @@ pub mod ops;
 pub mod path;
 pub mod project_repository;
 pub mod projects;
+pub mod rebase;
 pub mod remotes;
 pub mod ssh;
 pub mod storage;
